Add concept: "Mergable" to Pull Request
WIP (Work In Progress) - DO NOT MERGE

When Atlantis inspects a Pull Request to determine if it can
run "terraform apply" it (presently) only looks to see if any
of the reviewers have approved the PR.  This commit adds
the concept of "Mergeable" - affording projects the ability
to attract new contributors (by allowing them to approve
PR's) without allowing those PR's to be merged without
approval from the project owners.  The github v3 API
models "Mergeable" [1] for this purpose.

[1] https://godoc.org/github.com/google/go-github/github#PullRequest.GetMergeable

---

Remaining workitems:

- actually implement PullIsMergable()
- ensure coverage from e2e tests
- test with actual PR's from our terraform repo